% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{omega_range}
\alias{omega_range}
\title{omega_range}
\usage{
omega_range(low.logf, high.logf, p.per.dec)
}
\arguments{
\item{low.logf}{The log of the minimum frequency in Hz}

\item{high.logf}{The log of the maximum frequency in Hz}

\item{p.per.dec}{The number of points per decade.}
}
\description{
This function calculates a range of values for angular frequency (omega) from
specified limits of log frequency with a given number of points per decade of
frequency.
}
\examples{
omega.range(-1, 6, 10)
}
