% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{para}
\alias{para}
\title{para}
\usage{
para(...)
}
\arguments{
\item{...}{numeric or complex logical vectors passed as separate arguments}
}
\description{
This function returns the reciprocal of the sum of the reciprocals of all the
values present in its arguments.
}
\examples{
para(1, 2, 3, 4)
para(c(1 + 2i, 1 + 3i), c(2 + 1i, 1 - 3i), c(4 + 2i, 1 + 2i))

}
