% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{trans_line}
\alias{trans_line}
\title{trans_line}
\usage{
trans_line(R, U, n)
}
\arguments{
\item{R}{The resistance in one of the parallel rails per unit in Ohms.}

\item{U}{A complex vector (i.e., the result of any other impedance calculation)}

\item{n}{The number of units in the transmission line}
}
\description{
Calculates the impedance of a finite length transmission line where one rail has zero
impedance, the other parallel rail has resistance of R for each unit, where a component
U connects the two rails which is a complex vector of impedance values for any other
equivalent circuit, and the transmission line is n units long. The maximum length is 25
units.
}
\examples{
trans_line(R = 2, U = Z_C(1E-6, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10)), n  = 10) # this is equivalent to the FSW.
}
