% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_C}
\alias{Z_C}
\title{Z_C}
\usage{
Z_C(C, omega)
}
\arguments{
\item{C}{capacitance with units of Farads.}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of a capacitor from a capacitance and a vector of values of angular
frequency and returns a complex vector.
}
\examples{
Z_C(1E-4, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
