% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_FLG}
\alias{Z_FLG}
\title{Z_FLG}
\usage{
Z_FLG(Z0, k, tau, omega)
}
\arguments{
\item{Z0}{the DC resistance associated with the FLG element in Ohms}

\item{k}{the rate constant associated with the FLG element in seconds^-1}

\item{tau}{the time constant associated with the FLG element in seconds}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of a finite length Gerischer element from its Z0 value,
rate constant k, time constant tau and a vector of angular frequency
values
}
\examples{
Z_FLG(50, 100, 0.1, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
