% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_FLW}
\alias{Z_FLW}
\title{Z_FLW}
\usage{
Z_FLW(Z0, tau, omega)
}
\arguments{
\item{Z0}{the DC resistance associated with the FLW in Ohms.}

\item{tau}{the time constant of the FLW in seconds.}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of a finite length ("short") Warburg element from its Z0 value,
time constant tau and a vector of angular frequency values. Returns a complex vector.
}
\examples{
Z_FLW(50, 1, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
