% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_G}
\alias{Z_G}
\title{Z_G}
\usage{
Z_G(Z0, k, omega)
}
\arguments{
\item{Z0}{the DC resistance associated with the Gerischer element in Ohms}

\item{k}{the rate constant associated with the Gerischer element in seconds^-1}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of an infinite length Gerischer element from its Z0 value,
rate constant k and a vector of angular frequency values. Returns a complex vector.
}
\examples{
Z_G(50, 1, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
