% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_L}
\alias{Z_L}
\title{Z_L}
\usage{
Z_L(L, omega)
}
\arguments{
\item{L}{inductance with units of Henry.}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of an inductor from an inductance and a vector of values of angular
frequency and returns a complex vector.
}
\examples{
Z_L(1E-7, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
