% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_Q}
\alias{Z_Q}
\title{Z_Q}
\usage{
Z_Q(Q, n, omega)
}
\arguments{
\item{Q}{the "Q" value of the constant phase element (equal to C when n = 1), with units of Siemens * seconds^n.}

\item{n}{the "n" value where the phase angle is 90/n degrees.}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of a complex phase element from its "Q" value, the phase angle (90/n) degrees,
and a vector of angular frequency values. Returns a complex vector.
}
\examples{
Z_Q(1E-5, 0.8, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
