% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_R}
\alias{Z_R}
\title{Z_R}
\usage{
Z_R(R)
}
\arguments{
\item{R}{resistance with units of Ohms.}
}
\description{
Calculates the impedance of a resistor and returns a complex number (simply a real number with
a resistance of the argument R).
}
\examples{
Z_R(10)
}
