% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Z_to_df}
\alias{Z_to_df}
\title{Z_to_df}
\usage{
Z_to_df(Z)
}
\arguments{
\item{Z}{A complex vector.}
}
\description{
Takes a complex vector and returns a data frame of the real and imaginary parts
as separate columns.
}
\examples{
Z_to_df(para(Z_R(10), Z_C(1E-4, omega.range(-1, 6, 10))))
}
