% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circuit_elements.R
\name{Z_W}
\alias{Z_W}
\title{Z_W}
\usage{
Z_W(sigma, omega)
}
\arguments{
\item{sigma}{the Warburg coefficient, with units of Ohms * seconds^-1/2.}

\item{omega}{a vector of angular frequency values, which can be created with omega_range()}
}
\description{
Calculates the impedance of an infinite length Warburg element from the Warburg coefficient sigma
and a vector of angular frequency values. Returns a complex vector.
}
\examples{
Z_V(50, omega_range(low.logf = -1, high.logf = 5, p.per.dec = 10))
}
